@@ -1,6 +1,6 @@
 plugins {
     id 'java'
-    id 'org.springframework.boot' version '2.7.10'
+    id 'org.springframework.boot' version '3.0.5'
     id 'io.spring.dependency-management' version '1.0.11.RELEASE'
     id "org.asciidoctor.jvm.convert" version "3.3.2"
     id "com.ewerk.gradle.plugins.querydsl" version "1.0.10"
@@ -8,7 +8,7 @@ plugins {
 
 group = 'ybcharLog'
 version = '0.0.1-SNAPSHOT'
-sourceCompatibility = '11'
+sourceCompatibility = '17'
 
 configurations {
     compileOnly {
@@ -23,7 +23,6 @@ repositories {
 }
 
 ext {
-    asciidocVersion = "2.0.6.RELEASE"
     snippetsDir = file('build/generated-snippets')
     queryDslVersion = "5.0.0"
 }
@@ -54,25 +53,35 @@ dependencies {
     implementation 'org.springframework.security:spring-security-crypto'
     implementation 'org.bouncycastle:bcprov-jdk15on:1.70'
 
-    implementation 'com.querydsl:querydsl-core'
-    implementation 'com.querydsl:querydsl-jpa'
     implementation 'io.springfox:springfox-swagger-ui:3.0.0'
     implementation 'org.projectlombok:lombok'
     annotationProcessor('org.projectlombok:lombok')
+
 
     implementation 'org.projectlombok:lombok-mapstruct-binding:0.2.0'
     implementation 'org.mapstruct:mapstruct:1.5.3.Final'
     annotationProcessor 'org.projectlombok:lombok-mapstruct-binding:0.2.0'
     annotationProcessor 'org.mapstruct:mapstruct-processor:1.5.3.Final'
 
-    asciidoctorExt "org.springframework.restdocs:spring-restdocs-asciidoctor:${asciidocVersion}"
-    testImplementation "org.springframework.restdocs:spring-restdocs-mockmvc:${asciidocVersion}"
+    asciidoctorExt "org.springframework.restdocs:spring-restdocs-asciidoctor:3.0.0"
+    testImplementation "org.springframework.restdocs:spring-restdocs-mockmvc:3.0.0"
 
     // query dsl
     def queryDSL = '5.0.0'
 //    compile("com.querydsl:querydsl-jpa:${queryDSL}")
 //    compile("com.querydsl:querydsl-apt:${queryDSL}:jpa")
-    annotationProcessor("com.querydsl:querydsl-apt:${queryDSL}:jpa")
+//    implementation 'com.querydsl:querydsl-jpa'
+//    implementation 'com.querydsl:querydsl-core'
+//    annotationProcessor("com.querydsl:querydsl-apt:${queryDSL}:jpa")
+
+    implementation 'com.querydsl:querydsl-core:5.0.0'
+    implementation 'com.querydsl:querydsl-jpa:5.0.0:jakarta'
+
+    annotationProcessor "com.querydsl:querydsl-apt:5.0.0:jakarta"
+    annotationProcessor 'jakarta.persistence:jakarta.persistence-api'
+    annotationProcessor 'jakarta.annotation:jakarta.annotation-api'
+
+
     annotationProcessor("org.hibernate.javax.persistence:hibernate-jpa-2.1-api:1.0.2.Final")
     annotationProcessor("javax.annotation:javax.annotation-api:1.3.2")
 
