@@ -4,7 +4,7 @@ package com.ybcharlog.api.RequestDto.comment;
 import com.ybcharlog.api.domain.post.Post;
 import lombok.*;
 
-import javax.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotBlank;
 
 @Data
 @Getter
