@@ -2,7 +2,7 @@ package com.ybcharlog.api.RequestDto.comment;
 
 import lombok.*;
 
-import javax.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotBlank;
 
 @Getter
 @ToString
